Complete top to bottom refactor
Decided to completely re-do this project as the last iteration was a mess.

The mastermind.rb now just manages intro menus.
The standard_game.rb runs a standad game
code.rb is similar but simplified.
turn.rb is for tracking turn history
codemaker_game.rb is not yet implemented but will be for games where the user chooses the code and the computer guesses.